Moved the main script from main-v2.py (latest script) to main.py (prev unused script)
