+++ main.py
@@ -6,17 +6,17 @@ import numpy as np
 from datetime import datetime, date, time, timedelta
 from classes.timeslot import Timeslot
 from classes.proposal import Proposal
+import copy
 
-TIMESLOTS: list[Timeslot] = list()
 PROPOSALS: list[Proposal] = list()
-TIME_RESOLUTION: int = 60 * 60
-
+MIN_DATE: date = date.today()
+MAX_DATE: date = date.today()
 
 def read_proposals_from_csv(file_path: str) -> list[Proposal]:
     proposals = []
     with open(file_path, 'r') as file:
         reader = csv.DictReader(file)
-        for row in reader:
+        for id, row in enumerate(reader):
             prefered_dates_start = []
             prefered_dates_end = []
             avoid_dates_start = []
@@ -46,7 +46,8 @@ def read_proposals_from_csv(file_path: str) -> list[Proposal]:
             if row['minimum_antennas'] == '':
                 continue # invalid data with missing key 'minimum_antennas'
 
-
+            if int(row['simulated_duration']) < 60 * 30:
+                continue
             proposals.append(Proposal(
                 int(row['id']),
                 row['owner_email'],
@@ -55,15 +56,16 @@ def read_proposals_from_csv(file_path: str) -> list[Proposal]:
                 prefered_dates_end,
                 avoid_dates_start,
                 avoid_dates_end,
-                True if str(row['night_obs']).capitalize() == "Yes" else False,
-                True if str(row['avoid_sunrise_sunset']).capitalize == "Yes" else False,
+                True if str(row['night_obs']).lower() == "yes" else False,
+                True if str(row['avoid_sunrise_sunset']).lower == "yes" else False,
                 int(row['minimum_antennas']),
                 Proposal.parse_time(row['lst_start']),
                 Proposal.parse_time(row['lst_start_end']),
                 int(row['simulated_duration']),
                 get_score(str(row['proposal_id']))
             ))
-
+                
+    
     return proposals
 
 def get_score(proposal_id: str):
@@ -94,45 +96,14 @@ def generate_timeslots(start_date: date, end_date: date, timeslot_duration: int 
 #]
 
 
-def get_timeslot_by_id(timeslot_id: int) -> Timeslot:
-    global TIMESLOTS
-    return next((t for t in TIMESLOTS if t.id == timeslot_id), None)
 
 def get_proposal_by_id(proposal_id: int) -> Proposal:
     global PROPOSALS
     return next((p for p in PROPOSALS if p.id == proposal_id), None)
 
-def utc_to_lst(utc_datetime: datetime) -> datetime:
-    """
-    Convert UTC datetime to Local Sidereal Time (LST) datetime.
-    
-    Parameters:
-    utc_datetime (datetime): The UTC datetime object to be converted.
-    
-    Returns:
-    datetime: The corresponding LST datetime object.
-    """
-    # Get the longitude and latitude of the location (in degrees)
-    longitude = 18.478847  # Cape Town longitude
-    latitude = -33.944382  # Cape Town latitude
 
-    # Calculate the Sidereal Time at Greenwich (GST) in hours
-    gst_hours = (utc_datetime.timetuple().tm_year,
-                 utc_datetime.timetuple().tm_yday,
-                 utc_datetime.hour + utc_datetime.minute / 60 + utc_datetime.second / 3600)
-    gst_hours = (18.697374558 + 24.06570982441908 * (gst_hours[0] - 2000) + 0.000026 * (gst_hours[0] - 2000) ** 2 + 1.0027379093 * gst_hours[1] + gst_hours[2]) % 24
-
-    # Convert GST to LST
-    lst_hours = (gst_hours + longitude / 15) % 24
-
-    # Convert LST hours to LST datetime
-    lst_datetime = utc_datetime.replace(hour=int(lst_hours),
-                                       minute=int((lst_hours % 1) * 60),
-                                       second=int((lst_hours % 1) * 3600) % 60,
-                                       microsecond=0)
-
-    return lst_datetime
-
+def lst_to_utc(date: date, lst_time: time) -> datetime:
+    return datetime.combine(date, lst_time)
 
 def get_night_window(date: date) -> tuple[datetime, datetime]:
     """
@@ -167,292 +138,288 @@ def get_sunrise_sunset(date: date) -> tuple[datetime, datetime]:
     sunset_datetime = datetime(date.year, date.month, date.day, 18, 0, 0)  # 6:00 PM
     return sunrise_datetime, sunset_datetime
 
-
-
+# Checkout astropy to check lst start window, night window, and sunrise and sunset
 class Timetable:
-    def __init__(self, schedules: list[list[int]] = list()):
-        self.schedules = list()
-        if schedules == list():
+    def __init__(self, schedules: list[list[int]] = None):
+        if schedules is None:
+            self.schedules = list()
             self.generate()
         else:
             self.schedules = schedules
 
-    def generate(self) -> None:
-        global TIMESLOTS, PROPOSALS
-        for timeslot in TIMESLOTS:
-            proposal_id: int = None
-            if random.random() < 0.75:
-                for _ in range(6):
-                    proposal: Proposal = random.choice(PROPOSALS)
-                    if self.all_hard_contraints_met(timeslot, proposal):
-                        proposal_id = proposal.id
-                        break
-            self.schedules.append(
-                [timeslot.id, proposal_id]
-            )
-        return
-
-    def get_proposal_timeslot_indexes(self, proposal_id: int) -> list[int]:
-        timeslot_indexes: list[int] = []
-        for i, schedule in enumerate(self.schedules):
-            if schedule[1] == proposal_id:
-                timeslot_indexes.append(i)
-        return timeslot_indexes
+    def all_constraints_met(self, proposal: Proposal, start_datetime: datetime) -> bool:
+        
+        return self.night_obs_constraint_met(proposal, start_datetime) and self.avoid_sunrise_sunset_contraint_met(proposal, start_datetime)
     
-    def all_hard_contraints_met(self, timeslot: Timeslot, proposal: Proposal):
-        return self.is_avoid_sunrise_sunset_cosnstraint_met(timeslot, proposal) and self.is_night_obs_contraint_met(timeslot, proposal) and self.is_lst_start_start_end_contraint_met(timeslot, proposal) and self.is_over_schedule_contraint_met(timeslot, proposal)
+    def night_obs_constraint_met(self, proposal: Proposal, start_datetime: datetime) -> bool:
+        if proposal.night_obs:
+            # Compute end datetime based on simulated duration
+            end_datetime = start_datetime + timedelta(seconds=proposal.simulated_duration)
+            night_start_datetime, night_end_datetime = get_night_window(start_datetime.date())
 
-    def is_avoid_sunrise_sunset_cosnstraint_met(self, timeslot: Timeslot, proposal: Proposal):
+            # Check if both start and end datetimes fall within the night window
+            if night_start_datetime <= start_datetime <= night_end_datetime and \
+            night_start_datetime <= end_datetime <= night_end_datetime:
+                return True  # Constraint met
+            return False  # Constraint not met
+        return True  # If night observations are not required, constraint is met
+    
+    def avoid_sunrise_sunset_contraint_met(self, proposal: Proposal, start_datetime: datetime) -> bool:
         if proposal.avoid_sunrise_sunset:
-            sunrise_datetime, sunset_datetime = get_sunrise_sunset(timeslot.start_time.date())
-            
-            # Create a list of sunrise and sunset datetimes
-            sun_events = [sunrise_datetime, sunset_datetime]
+            # Get sunrise and sunset datetimes
+            sunrise_datetime, sunset_datetime = get_sunrise_sunset(date=start_datetime.date())
 
-            # Check if timeslot overlaps with any sun event
-            for event_datetime in sun_events:
-                if timeslot.start_time <= event_datetime <= timeslot.end_time:
-                    return False
-        return True
-    
-    def is_lst_start_start_end_contraint_met(self, timeslot: Timeslot, proposal: Proposal):
-        if len(self.schedules) == len(TIMESLOTS):
-            start_time: datetime = utc_to_lst(datetime.combine(timeslot.start_time.date(), proposal.lst_start_time))
-            start_end_time: datetime = utc_to_lst(datetime.combine(timeslot.start_time.date(), proposal.lst_start_end_time))
-            
-            timeslot_start_time: datetime = utc_to_lst(timeslot.start_time)
+            # Compute end datetime based on the proposal's duration
+            end_datetime = start_datetime + timedelta(minutes=proposal.simulated_duration)
 
-            timeslot_id = timeslot.id
-            proposal_id = proposal.id
+            # Check if sunrise or sunset occurs within the proposal's duration
+            if (sunrise_datetime >= end_datetime) or (sunset_datetime <= start_datetime):
+                return True  # Constraint met (neither occurs during the proposal)
+            return False  # Constraint not met (one of them occurs during the proposal)
+        return True  # If avoiding sunrise/sunset is not required, constraint is met
 
-            try:
-                timeslot_index: int = [schedule[0] for schedule in self.schedules].index(timeslot_id)
-                proposal_index: int = [schedule[1] for schedule in self.schedules].index(proposal_id)
-            except:
-                return True
-
-
-            if not (start_time <= timeslot_start_time and start_end_time >= timeslot_start_time) and (timeslot_index == proposal_index):
-                return False
-        return True
-    
-    def is_night_obs_contraint_met(self, timeslot: Timeslot, proposal: Proposal):
-        night_start_datetime, night_end_datetime = get_night_window(timeslot.start_time.date())
-        if proposal.night_obs and not (night_start_datetime <= timeslot.start_time <= night_end_datetime):
-            return False
-        return True
-    
-    def is_over_schedule_contraint_met(self, timeslot: Timeslot, proposal: Proposal):
-        num_of_timeslots: int = [schedule[1] for schedule in self.schedules].count(proposal.id)
-        if num_of_timeslots * timeslot.get_duration() >= proposal.simulated_duration:
-            return False
-        return True
-
-
-    def compute_penalty(self, proposal_id) -> float:
-        global TIME_RESOLUTION
-        penalty: float = 1
-        penalty_factors: list[float] = list([0.95, 0.90, 0.85, 0.80])
-        penalty_factor: float = penalty_factors[0]
-
-        scheduled_time: float = 0.0
-        requested_time: float = 0.0
-
-        """
-        List of contraints
-            - 1. Check for proposal score/priority SCI
-            - 2. Check LST start - start end window
-            - 3. Check for gaps between scheduled proposals
-            - 4. Check for partially allocated proposals
-            - 5. Check for night obs.
-            - 6. Check for avoid sunset/sunrise
-            - 7. Check for min antenna *
-        """
-        for schedule in self.schedules:
-            t_id, p_id = schedule
-            unique_proposal_ids: list[int] = list()
-            if p_id is not None and p_id == proposal_id:
-                timeslot = get_timeslot_by_id(t_id)
-                proposal = get_proposal_by_id(p_id)
-
-                # 1. Check for proposal score/priority
-                penalty_factor = penalty_factors[proposal.get_score() - 1]
-
-                # 2. Check LST start - start end window
-                start_time: datetime = utc_to_lst(datetime.combine(timeslot.start_time.date(), proposal.lst_start_time))
-                start_end_time: datetime = utc_to_lst(datetime.combine(timeslot.start_time.date(), proposal.lst_start_end_time))
-
-                timeslot_start_time: datetime = utc_to_lst(timeslot.start_time)
-                if not (start_time <= timeslot_start_time and start_end_time >= timeslot_start_time) and (proposal.id not in unique_proposal_ids):
-                    # Calculate the midpoint
-                    midpoint_start_time: datetime = start_time + (start_end_time - start_time) / 2
-                    
-                    #penalty = 0 # For hard contraints
-                    penalty *= (penalty_factor * 0.1) ** abs((midpoint_start_time - timeslot_start_time).total_seconds() / TIME_RESOLUTION)
-                    
-
-                # 3. Checking gaps between scheduled proposals
-                proposal_timeslot_indexes = self.get_proposal_timeslot_indexes(proposal_id)
-                if proposal.id not in unique_proposal_ids:
-                    penalty *= ((penalty_factor * 0.5) ** (abs((proposal.simulated_duration - timeslot.get_duration() * len(proposal_timeslot_indexes)) / TIME_RESOLUTION))) # Apply penalty for partially allocated proposal
-                
-                # 4. Checking partially allocated proposals
-                penalty *= (penalty_factor ** ((1 - (min(proposal_timeslot_indexes) + len(proposal_timeslot_indexes) - max(proposal_timeslot_indexes))) / TIME_RESOLUTION)) # Apply penalty for gaps of a partially allocated proposal
+    def generate_datetime(self, proposal_id: int) -> datetime:
+        proposal: Proposal = get_proposal_by_id(proposal_id)
         
+        for _ in range(10):  # Retry up to 10 times
+            # Randomly generate a date between MIN_DATE and MAX_DATE
+            randomly_generated_date = self.random_date(MIN_DATE, MAX_DATE)
+            
+            # Skip if the date exceeds MAX_DATE
+            if randomly_generated_date > MAX_DATE:
+                continue
 
-                # 5. Check for night obs.
-                night_start_datetime, night_end_datetime = get_night_window(timeslot.start_time.date())
+            # Randomly generate a time between lst_start_time and lst_end_time
+            randomly_generated_time = self.random_time(proposal.lst_start_time, proposal.lst_start_end_time)
 
-                # Check if the proposal requires night observation and if the timeslot is outside the night window
-                if proposal.night_obs and not (night_start_datetime <= timeslot.start_time <= night_end_datetime):
-                    # Calculate the midpoint of the night window
-                    night_midpoint_datetime = night_start_datetime + (night_end_datetime - night_start_datetime) / 2
+            # Combine date and time to get the start datetime
+            start_datetime = datetime.combine(randomly_generated_date, randomly_generated_time)
+            end_datetime = start_datetime + timedelta(seconds=proposal.simulated_duration)
+            #if start_datetime.weekday() == 3 or end_datetime.weekday() == 3: # Skip Wednsday
+            #    continue
+            
+            # Check if all constraints are met
+            if self.all_constraints_met(proposal, start_datetime):
+                return start_datetime
+        
+        return None  # Return None if no valid datetime is found after retries
+
+    def random_date(self, min_date: date, max_date: date) -> date:
+        delta = max_date - min_date
+        random_days = random.randint(0, delta.days)
+        return min_date + timedelta(days=random_days)
+
+    def random_time(self, start_time: time, end_time: time) -> time:
+        # Convert start and end time to total seconds
+        start_seconds = start_time.hour * 3600 + start_time.minute * 60 + start_time.second
+        end_seconds = end_time.hour * 3600 + end_time.minute * 60 + end_time.second
+
+        # If the end time is less than the start time, it means it wraps to the next day
+        if end_seconds < start_seconds:
+            end_seconds += 24 * 60 * 60  # Add 24 hours in seconds to end_time
+
+        # Generate a random time in seconds between start and end
+        random_seconds = random.randint(start_seconds, end_seconds)
+
+        # Convert back to hours, minutes, seconds
+        hours, remainder = divmod(random_seconds, 3600)
+        minutes, seconds = divmod(remainder, 60)
+        return time(hour=hours % 24, minute=minutes, second=seconds)  # Ensure hours are within 24
+
+    def generate(self) -> None:
+        global PROPOSALS
+        
+        for proposal in PROPOSALS:
+            start_datetime = self.generate_datetime(proposal.id) if random.random() > 0.75 else None
+            self.schedules.append([proposal.id, start_datetime])  # Append as a list
+        
+    
+    def compute_score(self) -> float:
+        global PROPOSALS
+        total_clash_time: float = 0
+        total_time: float = 0
+
+        for schedule in self.schedules:
+            proposal_id, start_datetime = schedule
+            if start_datetime is not None:
+                proposal: Proposal = get_proposal_by_id(proposal_id)
+                
+                
+                for another_proposal_id, another_proposal_start_datetime in self.schedules:
+                    another_proposal: Proposal = get_proposal_by_id(another_proposal_id)
                     
-                    # Calculate the midpoint of the timeslot
-                    timeslot_midpoint_datetime = timeslot.start_time + (timeslot.end_time - timeslot.start_time) / 2
-                    
-                    # Calculate the penalty based on the difference from the night midpoint
-                    penalty *= ((penalty_factor * 0.1) ** abs((timeslot_midpoint_datetime - night_midpoint_datetime).total_seconds() / TIME_RESOLUTION))
-                    #penalty = 0 # For hard contraints
+                    if (
+                        proposal_id != another_proposal_id and
+                        another_proposal_start_datetime is not None
+                    ):
+                        # Calculate end times for both proposals
+                        proposal_end_datetime = start_datetime + timedelta(seconds=proposal.simulated_duration)
+                        another_proposal_end_datetime = another_proposal_start_datetime + timedelta(seconds=another_proposal.simulated_duration)
 
-                    # 6. Check for avoid sunset/sunrise
-                    if proposal.avoid_sunrise_sunset:
-                        sunrise_datetime, sunset_datetime = get_sunrise_sunset(timeslot.start_time.date())
+                        # Check for clash
+                        # Calculate clash time
+                        max_start = max(start_datetime, another_proposal_start_datetime)
+                        min_end = min(proposal_end_datetime, another_proposal_end_datetime)
                         
-                        # Create a list of sunrise and sunset datetimes
-                        sun_events = [sunrise_datetime, sunset_datetime]
+                        # Calculate clash time in seconds
+                        clash_time = max(0, (min_end - max_start).total_seconds())
+                        total_clash_time += clash_time
 
-                        # Check if timeslot overlaps with any sun event
-                        for event_datetime in sun_events:
-                            if timeslot.start_time <= event_datetime <= timeslot.end_time:
-                                # Apply penalty based on the difference between the midpoint and the clashing event
-                                #penalty = 0 # For hard contraints
-                                penalty *= (penalty_factor * 0.1) ** abs((timeslot_midpoint_datetime - event_datetime).total_seconds() / TIME_RESOLUTION)
-                                break  # Exit loop after applying penalty for the first overlapping event
+                        total_time += proposal.simulated_duration
 
+        total_num_proposals = len(self.schedules)
+        total_num_unscheduled_proposals = [start_datetime for _, start_datetime in self.schedules].count(None)
 
+        # Simplified check for zero division
+        if total_time == 0 or total_num_proposals == 0:
+            return 0.0
 
-                # 7. Check for min antenna *
-                pass
+        numerator = 0
+        denominator = 1
 
-                
-                if proposal_id is not None:
-                    scheduled_time += timeslot.get_duration()
-                if proposal.id not in unique_proposal_ids:
-                    requested_time += proposal.simulated_duration
-                    unique_proposal_ids.append(proposal.id)
-        # 8. Compared scheduled used time with requestsed time
-        penalty *= scheduled_time / requested_time
+        total_non_clash_time = total_time - min(total_time * 0.5, total_clash_time * 25) # scalling up the clash time to increase it impact
+        total_num_scheduled_proposals = total_num_proposals - total_num_unscheduled_proposals
+        if total_time > total_num_proposals:
+            scaler = total_time / total_num_proposals
+            numerator = total_non_clash_time * 0.9 + scaler * total_num_scheduled_proposals * 0.1
+            denominator = total_time * 0.9 + scaler * total_num_proposals * 0.1
+        else:
+            scaler = total_num_proposals / total_time
+            numerator = scaler * total_non_clash_time * 0.9 + total_num_scheduled_proposals * 0.1
+            denominator = scaler * total_time * 0.9 + total_num_proposals * 0.1
+
+        score = numerator / denominator
+        # Calculate score
+        #score = ((total_time - total_clash_time) * (total_num_proposals - total_num_unscheduled_proposals) * 1.0) / (total_time * total_num_proposals * 1.0)
+        #score = ((total_time - total_clash_time) / (total_time))  * (0.9) ** ((total_num_proposals) / (total_num_proposals - total_num_unscheduled_proposals))
         
-        return penalty
-    
-    def score(self):
-        score: float = 0.0
-        #total_duration: float = 0.0
-        for schedule in self.schedules:
-            timeslot_id, proposal_id = schedule
-            if proposal_id is not None:
-                proposal = get_proposal_by_id(proposal_id)
-                #timeslot = get_timeslot_by_id(timeslot_id)
-                #score += proposal.score * self.compute_penalty(proposal_id)
-                score += self.compute_penalty(proposal_id)
-        return score #/ total_duration
+        # Ensure score is non-negative
+        return max(0.0, score)
 
-    
+            
     def crossover(self, schedules: list[list[int]]) -> list[list[int]]:
         offspring_schedules: list[list[int]] = list()
         for schedule_1, schedule_2 in zip(self.schedules, schedules):
             offspring_schedules.append(schedule_1 if random.random() > 0.5 else schedule_2)
         return offspring_schedules
     
-    def mutation(self, mutation_rate=0.2) -> None:
+    def mutation(self, mutation_rate=0.3) -> None:
         global PROPOSALS
         num_of_mutable_schedules: int = int(len(self.schedules) * mutation_rate)
-        for _ in range(num_of_mutable_schedules):
-            mutation_index = random.randint(0, len(self.schedules) - 1)
-            proposal_id: int = None
-            timeslot_id: int = self.schedules[mutation_index][0]
-            if random.random() < 0.75:
-                for _ in range(6):
-                    proposal: Proposal = random.choice(PROPOSALS)
-                    if self.all_hard_contraints_met(get_timeslot_by_id(timeslot_id), proposal):
-                        proposal_id = proposal.id
-            self.schedules[mutation_index] =  list([timeslot_id, proposal_id])
+        mutation_indexes: set[int] = set()  # Use a set for unique mutation indexes
 
-          
-    def remove_partialy_allocated_proposals(self) -> None:
+        # Create a deep copy of the schedules to avoid modifying the original
+        original_schedules = copy.deepcopy(self.schedules)
+
+        while len(mutation_indexes) < num_of_mutable_schedules:
+            mutation_index = random.randint(0, len(original_schedules) - 1)
+            
+            if mutation_index not in mutation_indexes:
+                proposal_id = original_schedules[mutation_index][0]  # Get proposal_id
+                start_datetime = self.generate_datetime(proposal_id) if random.random() > 0.75 else None  # Compute new start_datetime
+                
+                # Mutate the copied schedule
+                original_schedules[mutation_index][1] = start_datetime  # Mutate the start_datetime at this mutation index
+                mutation_indexes.add(mutation_index)
+
+        # Update self.schedules with the mutated schedules
+        self.schedules = original_schedules
+
+    
+    def remove_clashing_proposals(self) -> None:
         return
     
-    def display(self) -> None:
-        # Create a figure and axis
-        fig, ax = plt.subplots(figsize=(24, 12))
 
-        # Set the x-axis ticks to weekdays
-        weekdays = ['Sunday', 'Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday']
-        ax.set_xticks(np.arange(len(weekdays)) + 0.5)
-        ax.set_xticklabels(weekdays, ha='center')
-        ax.set_xlim(-0.5, len(weekdays) + 0.5)  # Add this line to ensure the last block is aligned
+    def plot(self, filename='weekly_timetable.png'):
+        # Define days of the week and colors
+        days_of_week = ['Sunday', 'Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday']
+        colors = ['gray', 'yellow', 'blue', 'red', 'green', 'brown', 'pink', 
+                  'lightgreen', 'lightblue', 'wheat', 'salmon', 'lightcoral', 'lightyellow']
 
-        # Set the y-axis ticks to timeslots
-        start_time = min(get_timeslot_by_id(timeslot_id).start_time for timeslot_id, _ in self.schedules)
-        end_time = max(get_timeslot_by_id(timeslot_id).end_time for timeslot_id, _ in self.schedules)
-        time_range = [start_time.replace(hour=h, minute=0) for h in range(24)]
-        ax.set_yticks(np.arange(len(time_range)))
-        ax.set_yticklabels([t.strftime('%H:%M') for t in time_range])
-        ax.set_ylim(0, 23 + 1)
+        # Create a figure for plotting
+        fig, ax = plt.subplots(figsize=(12, 8))
 
-        # Plot the scheduled proposals
-        owner_colors = {}
-        unique_proposals = []
-        for i, (timeslot_id, proposal_id) in enumerate(self.schedules):
-            timeslot: Timeslot = get_timeslot_by_id(timeslot_id)
-            proposal: Proposal = get_proposal_by_id(proposal_id)
-            if proposal:
-                # Check if proposal doesn't exist in our unique proposals list
-                if proposal.id not in unique_proposals:
-                    unique_proposals.append(proposal.id)
+        # Dictionary to hold color legend
+        legend_dict = {}
 
-                # Calculate the position of the proposal on the grid
-                weekday = timeslot.start_time.weekday()
-                time_index = timeslot.start_time.hour
+        # Total number of proposals
+        total_proposals = len(self.schedules)
+        scheduled_proposals = 0
 
-                # Determine the color based on the proposal ID
-                color = f'C{PROPOSALS.index(proposal)}'
+        # Iterate through schedules to plot each proposal
+        for idx, (proposal_id, start_datetime) in enumerate(self.schedules):
+            if start_datetime is not None:
+                scheduled_proposals += 1  # Count scheduled proposals
+                proposal: Proposal = get_proposal_by_id(proposal_id)
+                end_datetime = start_datetime + timedelta(seconds=proposal.simulated_duration)
 
-                # Add a rectangle for the proposal
-                rect = matplotlib.patches.Rectangle(((weekday + 1) % 7, time_index), 1, 1, facecolor=color, alpha=0.5, edgecolor='black', linewidth=1)
-                ax.add_patch(rect)
+                # Calculate the day of the week and time for plotting
+                day_index = start_datetime.weekday()  # Monday is 0 and Sunday is 6
+                start_time = start_datetime.hour + start_datetime.minute / 60.0
+                end_time = end_datetime.hour + end_datetime.minute / 60.0 
 
-               
-                ax.text((weekday + 1) % 7 + 0.5, time_index + 0.5, proposal.owner_email, ha='center', va='center', color='white')
+                # Randomly select a color for the block
+                color = colors[idx % len(colors)]
 
-        # Add a legend
-        #legend_patches = [plt.Rectangle((0, 0), 1, 1, facecolor=f'C{i}', alpha=0.5, edgecolor='black', linewidth=2) for i in range(len(unique_proposals))]
-        #legend_labels = [f'Proposal {i}' for i in unique_proposals]
-        #ax.legend(legend_patches, legend_labels, loc='upper left', bbox_to_anchor=(1.05, 1))
+                # Handle overnight events
+                if (end_time - start_time) < 0:
+                    # Draw the first rectangle for the current day
+                    ax.fill_between([day_index - 0.5, day_index + 0.5], [start_time, start_time], [24, 24],
+                                    color=color, edgecolor='black', linewidth=0.5, alpha=0.25)
+                    # Draw the second rectangle for the next day
+                    next_day_index = (day_index + 1)  # Wrap around to the start of the week
+                    next_start_time = 0  # Start at the bottom of the next day
+                    if next_day_index < 7:
+                        ax.fill_between([next_day_index - 0.5, next_day_index + 0.5], [next_start_time, next_start_time], [end_time, end_time],
+                                        color=color, edgecolor='black', linewidth=0.5, alpha=0.25)
 
-        # Set the title and axis labels
-        ax.set_title('Timetable')
-        ax.set_xlabel('Weekday')
-        ax.set_ylabel('Time')
+                    # Place index text in the first rectangle
+                    ax.text(day_index, (start_time + 24) / 2, str(idx), 
+                            ha='center', va='center', fontsize=10, color='black')
+                    if next_day_index < 7:
+                        # Place index text in the second rectangle
+                        ax.text(next_day_index, (next_start_time + end_time) / 2, str(idx), 
+                                ha='center', va='center', fontsize=10, color='black')
 
-        # Display the plot
-        plt.savefig(f"outputs/timetable_{datetime.now()}.png")
+                else:
+                    # Plot the block for the proposal with a black border and 25% opacity
+                    ax.fill_between([day_index - 0.5, day_index + 0.5], [start_time, start_time], [end_time, end_time],
+                                    color=color, edgecolor='black', linewidth=0.5, alpha=0.25)
 
-        # Print the textual output
-        for timeslot_id, proposal_id in self.schedules:
-            timeslot = get_timeslot_by_id(timeslot_id)
-            proposal = get_proposal_by_id(proposal_id)
-            if timeslot.start_time.hour == 0:
-                print("--------------------------------")
-                print(f"{timeslot.start_time.strftime('%d %B %Y')}")
-            print(f"\t{timeslot.start_time.strftime('%H:%M')} - {timeslot.end_time.strftime('%H:%M')}\t", end="")
-            if proposal:
-                print(proposal.owner_email, proposal.simulated_duration // (60 * 60))
-            else:
-                print("")
+                    # Place index text inside the rectangle
+                    ax.text(day_index, (start_time + end_time) / 2, str(idx), 
+                            ha='center', va='center', fontsize=10, color='black')
 
-        print(f"Fitness: {self.score():.2f}\n")
+                # Add to legend with index and email
+                legend_key = f'{idx} {proposal.owner_email.split("@")[0].split(".")[0]} {proposal.lst_start_time.strftime("%H:%M:%S")} {proposal.lst_start_end_time.strftime("%H:%M:%S")} {proposal.night_obs} {proposal.avoid_sunrise_sunset}'
+                if legend_key not in legend_dict:
+                    legend_dict[legend_key] = color
+
+                print(f"{idx}\t{day_index}\t{start_time:0.2f}\t{end_time:0.2f}\t\t{(end_time - start_time):0.2f}")
+     
+
+        # Set the x-axis and y-axis limits and labels
+        ax.set_xticks(range(len(days_of_week)))
+        ax.set_xticklabels(days_of_week)
+        ax.set_xlabel('Days of the Week')
+        ax.set_ylabel('Time (hours)')
+        ax.set_ylim(0, 24)
+        
+        # Update the title to include scheduled proposals count
+        ax.set_title(f'Weekly Timetable: {scheduled_proposals}/{total_proposals} Proposals', fontsize=16)
+
+        # Add gridlines for better readability
+        ax.yaxis.grid(True)
+
+        # Create a legend outside the plot
+        handles = [plt.Rectangle((0, 0), 1, 1, color=color) for color in legend_dict.values()]
+        ax.legend(handles, legend_dict.keys(), title="Proposals", loc='upper left', bbox_to_anchor=(1, 1))
+
+        # Save the plot to a file
+        plt.tight_layout()
+        plt.savefig(filename, dpi=200)  # Save the figure as a PNG file
+        plt.close(fig)  # Close the figure to free up memory
+
 
 
 class GeneticAlgorithm():
@@ -463,7 +430,7 @@ class GeneticAlgorithm():
         self.generate_timetables()
 
         for generation in range(num_of_generations):
-            self.timetables.sort(key=lambda timetable: timetable.score(), reverse=True)
+            self.timetables.sort(key=lambda timetable: timetable.compute_score(), reverse=True)
             self.print_fitness(generation)
             self.evolve()
             
@@ -474,9 +441,9 @@ class GeneticAlgorithm():
 
     def evolve(self, crossover_rate: float = 0.2, mutation_rate: float = 0.1) -> None:
         # Elitism: Keep the best timetables
-        self.timetables.sort(key=lambda timetable: timetable.score(), reverse=True)
+        self.timetables.sort(key=lambda timetable: timetable.compute_score(), reverse=True)
         elite_timetables: list[Timetable] = self.timetables[:int(self.num_of_timetables * 0.75)]
-        #print(self.timetables[0].score(), self.timetables[-1].score(), end="\n\n")
+        #print(self.timetables[0].compute_score(), self.timetables[-1].compute_score(), end="\n\n")
         
         
         """
@@ -491,7 +458,7 @@ class GeneticAlgorithm():
 
         self.timetables.clear()
         self.timetables.append(elite_timetables[0])
-        print(self.timetables[0].score())
+        print(self.timetables[0].compute_score())
         # Replace the old population with the new one
         for timetable in new_timetables:
             self.timetables.append(timetable)
@@ -514,7 +481,7 @@ class GeneticAlgorithm():
                 offspring: Timetable = Timetable(parent_timetable_1.crossover(parent_timetable_2.schedules))
                 offspring.mutation(mutation_rate=0.2)
                 offsprings.append(offspring)
-            offsprings.sort(key=lambda timetable: timetable.score(), reverse=True)
+            offsprings.sort(key=lambda timetable: timetable.compute_score(), reverse=True)
             offsprint_timetable: Timetable = random.choice(offsprings[:max(2, int(num_offsprings * 0.4))])
             #offsprint_timetable.mutation()
             self.timetables[index] = offsprint_timetable
@@ -529,7 +496,7 @@ class GeneticAlgorithm():
 
     def print_fitness(self, generation: int) -> None:
         # Calculate fitness scores for the first up to 5 timetables
-        fitness_scores = [timetable.score() for timetable in self.timetables[:min(5, len(self.timetables))]]
+        fitness_scores = [timetable.compute_score() for timetable in self.timetables[:min(5, len(self.timetables))]]
         
         # Sort fitness scores in descending order
         sorted_scores = sorted(fitness_scores, reverse=True)
@@ -547,50 +514,92 @@ class GeneticAlgorithm():
 
     def get_best_fit_timetable(self) -> Timetable:
         # Sort timetables by score and return the best one
-        self.timetables.sort(key=lambda timetable: timetable.score(), reverse=True)
-        self.timetables[0].remove_partialy_allocated_proposals()
+        self.timetables.sort(key=lambda timetable: timetable.compute_score(), reverse=True)
+        self.timetables[0].remove_clashing_proposals()
         return self.timetables[0]
+    
+def can_schedule_proposal(proposal: Proposal) -> bool:
+    """Check if a proposal can be scheduled based on its constraints."""
+    for day in range((MAX_DATE - MIN_DATE).days + 1):  # Include the last day
+        # Get the night window for the current day
+        night_start_datetime, night_end_datetime = get_night_window(MIN_DATE + timedelta(days=day))
+        sunrise_datetime, sunset_datetime = get_sunrise_sunset(MIN_DATE + timedelta(days=day))
+
+        # Prepare the start times
+        min_start_datetime = lst_to_utc(MIN_DATE + timedelta(days=day), proposal.lst_start_time)
+        max_start_datetime = lst_to_utc(MIN_DATE + timedelta(days=day), proposal.lst_start_end_time)
+
+        # Check both start times for night observations
+        for start_datetime in [min_start_datetime, max_start_datetime]:
+            # Check if scheduling within the night window is possible
+            if proposal.night_obs:
+                if not (start_datetime >= night_start_datetime and 
+                        start_datetime + timedelta(seconds=proposal.simulated_duration) <= night_end_datetime):
+                    return False  # Constraint not met
+
+            # Check for sunrise/sunset avoidance
+            if proposal.avoid_sunrise_sunset:
+                if not (start_datetime + timedelta(seconds=proposal.simulated_duration) <= sunrise_datetime or 
+                        start_datetime >= sunset_datetime):
+                    return False  # Constraint not met
+
+    return True  # All constraints met
+
+
 
 
 def main():
-    global TIMESLOTS, PROPOSALS
+    global PROPOSALS, MIN_DATE, MAX_DATE
+    MIN_DATE = date(2025, 2, 9)
+    MAX_DATE = date(2025, 2, 15)
+    #proposals: list[Proposal] = read_proposals_from_csv('./proposals/csv/ProdObsList1743669829782.csv')
     proposals: list[Proposal] = read_proposals_from_csv('./proposals/csv/ObsList1737538994939.csv')
-    total_week_duration: int = 60 * 60 * 24
+    random.shuffle(proposals) # Shaffle the proposals
+
+    total_week_duration: int = (MAX_DATE - MIN_DATE - timedelta(days=1)).total_seconds()
     cumulative_week_duration: int = 0
     for proposal in proposals:
+        # Check if the proposal can be scheduled
+        if not can_schedule_proposal(proposal):
+            print("**", proposal.owner_email, proposal.night_obs, proposal.avoid_sunrise_sunset, proposal.lst_start_time, proposal.lst_start_end_time)
+            continue  # Skip this proposal if it cannot be scheduled
+        if proposal.night_obs or proposal.avoid_sunrise_sunset:
+            print(proposal.owner_email)
+        
         cumulative_week_duration += proposal.simulated_duration
-        if cumulative_week_duration >= total_week_duration * 2:
+        if cumulative_week_duration > total_week_duration * 0.85:
             break
+
+        # If the proposal is valid, add it to the scheduled proposals
         PROPOSALS.append(proposal)
-    TIMESLOTS = generate_timeslots(date.fromisoformat("2025-02-09"), date.fromisoformat("2025-02-15"))
 
     """
     print("Desplaying Randomly Generated Timetable...")
     timetable = Timetable()
-    timetable.display()
+    timetable.plot(filename=f'outputs/week timetable')
 
     print("Crossover...")
     parent_timetable_1: Timetable = Timetable()
     parent_timetable_2: Timetable = Timetable()
-    print(f"parent_timetable_1 fitnest {parent_timetable_1.score()}\tparent_timetable_2 fitnest {parent_timetable_2.score()}")
+    print(f"parent_timetable_1 fitnest {parent_timetable_1.compute_score()}\tparent_timetable_2 fitnest {parent_timetable_2.compute_score()}")
     offspring_timetable: Timetable = Timetable(parent_timetable_1.crossover(parent_timetable_2.schedules))
-    print(f"offspring_timetable fitness {offspring_timetable.score()}")
+    print(f"offspring_timetable fitness {offspring_timetable.compute_score()}")
 
     print("Mutation...")
     offspring_timetable.mutation()
-    print(f"offspring_timetable fitness after mutation {offspring_timetable.score()}")
+    print(f"offspring_timetable fitness after mutation {offspring_timetable.compute_score()}")
 
-    print(f"offspring_timetable fitnest {offspring_timetable.score()}\tparent_timetable_2 fitnest {parent_timetable_2.score()}")
+    print(f"offspring_timetable fitnest {offspring_timetable.compute_score()}\tparent_timetable_2 fitnest {parent_timetable_2.compute_score()}")
     offspring_timetable_1: Timetable = Timetable(offspring_timetable.crossover(parent_timetable_2.schedules))
-    print(f"offspring_timetable_1 fitness {offspring_timetable_1.score()}")
+    print(f"offspring_timetable_1 fitness {offspring_timetable_1.compute_score()}")
     
     exit()
     """
 
     print("Generating Timetable using Genetic Algorithim")
-    genetic_algorithm: GeneticAlgorithm = GeneticAlgorithm(20, 10000)
+    genetic_algorithm: GeneticAlgorithm = GeneticAlgorithm(10, 5000)
     best_timetable: Timetable = genetic_algorithm.get_best_fit_timetable()
-    best_timetable.display()
-
+    best_timetable.plot(filename=f'outputs/week {MIN_DATE.strftime("%m-%d-%Y")} to {MAX_DATE.strftime("%m-%d-%Y")} timetable.png')
+    
 if __name__ == "__main__":
     main()
